pyproject.toml
[tool.poetry]
name = "google-adk-tool-learning"
version = "1.1.0"
description = "Google ADK Toolkit"
authors = ["Kepler <austin.noronha@accrete.ai>"]
readme = "README.md"
packages = [{include = "google_adk_tool"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4"
requests = ">=2.31.0"
python-dateutil = ">=2.8.0"
bcrypt = ">=5.0.0"
python-dotenv = ">=1.0.0"
pre-commit = "^4.5.0"
google-adk = {extras = ["extensions"], version = "^1.25.0"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.2"
black = "^24.10.0"
isort = "^5.13.2"
ssort = "^0.13.0"
pytest = "8.3.1"
pytest-cov = "5.0.0"
pytest-mock = "3.14.0"
pytest-asyncio = "^1.3.0"
coverage = "7.6.1"
pre-commit = "^4.0.1"
bump-pydantic = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 150
target-version = ["py38", "py39"]
skip-string-normalization = true

[tool.ruff]
line-length = 150 
indent-width = 4
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]


[tool.ruff.lint]
extend-select = [
    # "D",     # pydocstyle
    "E501",  # enforce line length
    # "UP",    # pyupgrade
]

ignore = [
    "E722", # Do not use bare `except`
    "E402", 
    "F841", 
    "E711", 
    "E721", 
    "E712",
    "E501",
    "F811",
    "F601"
]

[tool.isort]
profile = "black"
line_length = 79

[tool.ssort]
# Enable grouping imports
group_imports = true

[tool.pytest.ini_options]
#addopts = "tests -vv --cov --cov-config .coveragerc"
# --cov-report html:coverage_html
# --cov-report xml:coverage.xml
addopts = """
    --cov=development_tutor
    --cov-report term-missing
"""
testpaths = [
    "tests",
]
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
# Code under development_tutor is what tests import; measure that.
source = ["development_tutor"]
# Optional: Omit specific files or directories from measurement.
# This helps ensure test files are not counted, even if in a subdirectory
# that might be imported by the source code.
omit = ["tests/*", "_backup/*", "__init__.py"]

[tool.coverage.report]
# Optional: Omit specific files from the final report.
# This is an additional safeguard.
omit = ["tests/*", "_backup/*", "__init__.py"]
#fail_under = 70
show_missing = true


